ml-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
boto3==1.28.57
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
kagglehub==0.2.5
xgboost==1.7.6
lightgbm==4.0.0
shap==0.43.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
jupyter==1.0.0